+++ edgar-app/src/components/navigation/ResponsiveNavBar.tsx
